Participants_Of_Game
This program asks the user to enter the names of participants in a game and their subsequent scores. The program then sorts these scores in descending order through a bubble sort. The final result the prints the final results table where the participant with the highest score will be on top and the lowest scoring participant will sit at the bottom.