Practising BubbleSort programs
This program is a quick practice of my Bubble Sort algorithms where the user enters the number of elements they want into an array then input the values of the elements themselves as doubles. We then call the BubbleSort method which sorts the inputs into ascending order (smallest to the top, largest to the bottom). The program then prints both the original and sorted array in comparison.